requirements.txt
# --- UI & GRAPHICS ---
customtkinter==5.2.2
pillow==12.1.0

# --- DATABASES (PART 3) ---
pymongo==4.6.1
PyMySQL==1.1.2
mysql-connector-python==8.2.0

# --- TRANSLATION & RTL (PART 4) ---
deep-translator==1.11.4
arabic-reshaper==3.0.0
python-bidi==0.4.2
requests==2.32.5

# --- TESTING & BEST PRACTICES (PART 5) ---
pytest==9.0.2
pytest-mock==3.15.1
mock==5.1.0
cryptography==42.0.2